--- jsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": false,
-    "target": "es2017",
-    "allowSyntheticDefaultImports": false,
-    "baseUrl": "./src",
-    "paths": {
-      "src/*": ["src/*"],
-      "components/*": ["src/components/*"],
-      "services/*": ["src/services/*"],
-      "uikit/*": ["src/uikit/*"],
-      "assets/*": ["src/assets/*"],
-      "common/*": ["src/common/*"],
-      "shapes/*": ["src/shapes/*"],
-      "store/*": ["src/store/*"]
-    }
-  },
-  "exclude": ["node_modules", "dist"]
-}
