+++ package.json
@@ -8,7 +8,7 @@
     "@nivo/core": "^0.51.0",
     "@nivo/pie": "^0.51.0",
     "@oncojs/survivalplot": "^0.7.2",
-    "antd": "^3.25.3",
+    "antd": "^3.26.12",
     "auto-bind-es5": "^1.1.0",
     "axios": "^0.19.0",
     "file-saver": "^2.0.2",
@@ -82,15 +82,15 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.3",
     "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.28",
+    "@types/jest": "^25.1.3",
+    "@types/node": "^13.7.7",
     "@types/react": "^16.9.23",
     "@types/react-dom": "^16.9.5",
     "chai": "^4.2.0",
     "less": "^3.10.3",
     "mocha": "^6.2.2",
     "prettier": "1.19.1",
-    "source-map-explorer": "^2.1.2",
-    "typescript": "^3.8.2"
+    "source-map-explorer": "^2.3.1",
+    "typescript": "^3.8.3"
   }
 }
