+++ src/components/CohortBuilder/ModalContentSection.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+type ModalContentSectionProps = {
+  children: object;
+};
+
+export const ModalContentSection = (props: ModalContentSectionProps) => (
+  <div className="cb-modalContentSection">{props.children}</div>
+);
